reorg diffmap
